+++ Assets/Scripts/Animat/Brain/GPU/Resources/CPGClass.cs
@@ -24,19 +24,19 @@ public class CPG{
     this.r_s = r_s;
     n = neuron_set.Length;
   }
-  private float Sigmoid(float x){
-    return (float)(1.0 /(1.0 + Math.Exp(-x)));
+  private double Sigmoid(double x){
+    return (double)(1.0 /(1.0 + Math.Exp(-x)));
   }
   public void updateActivations(){
     for(int i = 0; i < n; i++){
-      neuron_set[i].activation = (float)(r_s[i]*Math.Sin(w_s[i]*(Time.time - p_s[i])) + (1 - r_s[i])*Sigmoid(neuron_set[i].activation));
+      neuron_set[i].activation = (double)(r_s[i]*Math.Sin(w_s[i]*(Time.time - p_s[i])) + (1 - r_s[i])*Sigmoid(neuron_set[i].activation));
     }
   }
-  public float[] fitnessFunction(){
+  public double[] fitnessFunction(){
     int num_per_limb = neuron_set.Length/4;
-    float[] fitnessPerLimb = new float[4];
+    double[] fitnessPerLimb = new double[4];
     fitnessPerLimb[0] = 0; fitnessPerLimb[1] = 0; fitnessPerLimb[2] = 0; fitnessPerLimb[3] = 0;
-    float min_act = 1000000000;
+    double min_act = 1000000000;
     for(int i = 0; i < num_per_limb; i++){
       if(neuron_set[i].activation < min_act){
         min_act = neuron_set[i].activation;
